requirements.txt
streamlit
pandas
matplotlib
easyocr
Pillow
numpy
